refactor(page): ShowPage rename
Renaming a ShowPage to a DetailPage
