+++ src/Pages/Crud/DetailPage.php
@@ -4,7 +4,7 @@ namespace MoonShine\Pages\Crud;
 
 use Illuminate\View\ComponentAttributeBag;
 use MoonShine\Buttons\IndexPage\DeleteButton;
-use MoonShine\Buttons\ShowPage\FormButton;
+use MoonShine\Buttons\DetailPage\FormButton;
 use MoonShine\Components\ActionGroup;
 use MoonShine\Components\TableBuilder;
 use MoonShine\Decorations\Block;
@@ -14,7 +14,7 @@ use MoonShine\Decorations\Fragment;
 use MoonShine\Pages\Page;
 use Throwable;
 
-class ShowPage extends Page
+class DetailPage extends Page
 {
     public function breadcrumbs(): array
     {
